Datetime extensions

+++ sources/core-domain/src/main/java/com/sedsoftware/core/domain/type/RemindiePeriod.kt
@@ -1,12 +1,12 @@
 package com.sedsoftware.core.domain.type
 
-sealed class RemindiePeriod(val each: Int) {
-    object None : RemindiePeriod(0)
-    class Hourly(each: Int) : RemindiePeriod(each)
-    class Daily(each: Int) : RemindiePeriod(each)
-    class Weekly(each: Int) : RemindiePeriod(each)
-    class Monthly(each: Int) : RemindiePeriod(each)
-    class Yearly(each: Int) : RemindiePeriod(each)
+sealed class RemindiePeriod(val each: Long) {
+    object None : RemindiePeriod(0L)
+    class Hourly(each: Long) : RemindiePeriod(each)
+    class Daily(each: Long) : RemindiePeriod(each)
+    class Weekly(each: Long) : RemindiePeriod(each)
+    class Monthly(each: Long) : RemindiePeriod(each)
+    class Yearly(each: Long) : RemindiePeriod(each)
 
     override fun toString(): String = "${this.javaClass.simpleName} [$each]"
 }
